Title: 최대 숫자 만들기, Time: 12ms, Memory: 0MB, Status: Runtime Error - CodeTree

+++ 240130/최대 숫자 만들기/make-biggest-num.cpp
@@ -8,8 +8,8 @@ using namespace std;
 bool cmp(string a, string b){
     string str1 = a + b;
     string str2 = b + a;
-    int num1 = stoi(str1);
-    int num2 = stoi(str2);
+    long long num1 = stoi(str1);
+    long long num2 = stoi(str2);
     if(num1 > num2){
         return true;
     }
